Add the invert buttons
controlgroup.h
--------------
Just made cosmetic changes for now. Need to perfect the automated layout,
but the automated connections to signal mapper worked perfectly. Would
like to use a QList of pointers to the caller's slot function(s), but there
are bigger fish to fry right now.

bits.cpp, bits.h, bits.ui.h
---------------------------
Added support for the invert buttons and did a little comment cleanup.

Still need to do work on the layout and autosizing of the ui.