Finished support code for invert buttons
Used the ControlGroup template class to create the invert buttons for the
HexEdit combo boxes. These buttons simply have the til ~ for text, since
every C programmer knows that this is the NOT of the number.

bits.cpp
--------
* Changed the argument in the call to updateHexEdit to quint64, rather than
  void*. This is simpler and actually minimizes type conversions.

* Added a parameter to the showDecimals() function. It was getting its values
  from counting the set bits, but it is not always the set bits that change.
  Of course, you could require it to be called in a sequence that guarantees
  that the bits have been updated, but this won't be true unless the bits
  are connected to the HexEdit box that is being updated. And I wanted to
  display the decimal equivalents in the message area from any changes in
  any of the HexEdit boxes.

* onInvert slot function was changed to make calls into the HexEdit class
  for hexstring and hexvalue updates. This centralizes the code to update
  these number and string objects, thereby minimizing cruft.

* In updateBits(), removed the cruft that could now be replaced by the
  functionality centralized in the HexEdit class.

* showDecimals() had a bug where 64-bit positive numbers were being shown in
  red, indicating that they were incorrectly interpreted as being negative.
  This was due to this operation "(1 << (bits-1)), when bits-1 evaluates to
  63. What happens is that the 1 is taken for a plain signed int, so
  shifiting it up that far made a 40000000.00000000 look like
  FFFFFFFF.80000000, so it was always interpreted as negative. The fix was
  to cast the nubmer 1 as quint64 like this: (quint64)1 << (bits-1>. I've
  run into this problem before, so it did not take long to find it.

* Changed the Twidget structure, so I had to change the way it is
  initialized in init_bitSizes() and init_invert(). These changes make the
  layout much more flexible and manageable.

bits.h
------
* Changed the prototype for showDecimals to pass the quint64 argument.

controlgroup.h
--------------
This file declars and defines the Twidget class and the ControlGroup class
template. The reason for the template is that we need to be able to
instantiate any kind of QWidget type, from QRadioButtons to QPushButtons to
QComboBoxes, QLabels, etc.

I needed a better, more reliable, and less cryptic way in which to layout
objects created by this class.

The Twidget class serves as an interface between the ControlGroup and its
callers, so that interface had to be changed.

I was laying out the objects algorithmically, but laying out tables like
that was very difficult, and the code was very obscure. The solution was to
simply use a QList of QPoints to locate each object and a QList of QSize to
size them. This is much more flexible, because now the objects are placed
absolutely, rather than relatively. That means each object can be anywhere
in the layout, and the objects can have different sizes.

hexedit.cpp
-----------
* updateHexEdit() - Changed the argument from void* to quint64 in order to
  simplify the interface. Had to change all the calls, too. Makes a good
  reason for function templates.

* hexstr2int()
  Moved the code that trims and strips the '.' out of the hexstring out of
  the bits class to this class and function, where it belongs. More
  centralization to limit the spread of cruft.