Fixed a couple bugs and added a little more functionality.
One bug updated all three combo boxes with the same data when changing the
bitsize. Change was made in HexEdit.cpp to update the currentText of the
window with the new bitfield width, not with the data in the BitButtons.

Added some commentary to the rather cryptic code in HexEdit.

Changed the setGeometry calls to use relative values as much as possible, so
that any changes in window geometry would be updated for all child objects,
rather than having to edit each one separately.