Got the format working with validation of input.
Still have work to do on the ouput.

Need to know the correct timing for setting and unsetting the inputMask
for the Result box. If the inputMask is set to hhhhhhhh.hhhhhhhh like
all the others, the format gets munged. If I elminate inputMas for the
results box altogether, then 64-bit output will not have the nice '.'
delimiter at 32 bits.

I'll need to think about this a bit yet.