bits.cpp -------- onShift() onCalc() keyPressEvent()
In all the above functions, a call to onFormat() was made to format
the result box if there was a format in the format box.

hexedit.cpp
-----------
Insert at item 1 seems to fix the problem where the read of the
current text in a hexEdit box is stale for some reason. See the
"To Do" in bits.cpp-onFormat.

shiftop.cpp
-----------
Changed the timing of the autorepeat for the shift buttons, but it
didn't seem to help the double bounce for the right button. Left
button seems to work fine.